Remove leading _ and collapse multiple consecutive _
This should result in slightly nicer names, for example when the original
file name already contained _ characters and adjacent characters have
been replaced. This comes at the price of having more collisions if
multiple files get mapped to the same sanitized name though.

Also extend the tests with these new cases and make use of positive look
behind and look ahead to avoid an empty string as the resulting name
in some edge cases when removing leading and trailing _.