Include weekday name in directory names by default
People might better remember the name of the current weekday than
the current date, so this might make the names a bit more useful.
Also make the pattern configurable and localizable to increase
flexibility.

Also change the code formatting in the tests a bit to better
visualise the directory structure that is expected and to shorten
some overly long lines a bit.

This basically implements the functionality suggested in GitHub
issue #4 [1].

[1]: https://github.com/mstock/RoomAssets-App-Sync/issues/4