Drop `--detailed-exitcode` parameter again
Implement a behaviour which exits with 0 on success if there were no
changes, 1 on success if there were changes and >= 2 if there was an
error instead. This should be a bit easier to use in shell scripts and
was suggested in GitHub issue #5 [1]. Also document this behaviour in
README.md.

[1]: https://github.com/mstock/RoomAssets-App-Sync/issues/5