Decode URI escaped values including umlauts in resource URLs
So far, the filename was extracted from the resource URL in a too simple
way, i.e. URI escaped characters where not unescaped, resulting in
percent encoded parts in the resulting file names. This should be fixed
now by retrieving the last path segment of the resource URL object and
using that as the file name and should thus fix GitHub issue #8 [1].

[1]: https://github.com/mstock/RoomAssets-App-Sync/issues/8