Drop id from talk directories, use only code
Apparently, the id does not identify a talk slot, but something else,
maybe a talk slot in a specific schedule version. This should thus fix
GitHub issue #9 [1]. Unfortunately, this is complicated a bit by the
fact that a talk can have multiple slots which share the same code, so
the logic which moves around existing talk directories if e.g. a talk
title was changed or the talk was moved around in the schedule also had
to be adapted to make it capable of handling this.

[1]: https://github.com/mstock/RoomAssets-App-Sync/issues/9