Skip files without suffix
These tend to be not directly usable on some operating systems and often
are the result of links to external resources like GitHub projects.
Especially in the latter case, the downloaded data also tends to change
on each update and would then be uploaded to Nextcloud again,
potentially resulting in a notification email about the update.