Respect umask when creating state file
File::Temp defaults to 0600 for the temporary file. However, it might be
useful to have a state file with permissions that are influenced by the
user's umask, just like it is the case for all the other files and
directories that are being created, such that e.g. a monitoring system
or backup job that use a different user may access it without elevating
their privileges.